python_server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Ben-Hu <benjqh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.32.0"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.32.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
